# lib/api-zod/src/generated/types/packInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackInputLanguage } from "./packInputLanguage";

export interface PackInput {
  /** @minLength 1 */
  name: string;
  language: PackInputLanguage;
}