# lib/api-zod/src/generated/types/packWithCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PackWithCount {
  id: number;
  userId: number;
  name: string;
  language: string;
  learned: number;
  wordCount: number;
}