# lib/api-zod/src/generated/types/userWithCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserWithCount {
  id: number;
  name: string;
  packCount: number;
}