# lib/api-zod/src/generated/types/wordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordInput {
  /** @minLength 1 */
  term: string;
  /** @minLength 1 */
  meaning: string;
}